+++ log.txt
@@ -1,2 +1,3 @@
 [2021-01-04  0:02:54] 잔디를 심었습니다.
 [2021-01-04  0:04:02] 잔디를 심었습니다. 
+[2021-01-04  0:15:24] 잔디를 심었습니다.
