--- src/main/java/com/junshock/jpatest/MemberRepository.java
@@ -1,23 +1,0 @@
-package com.junshock.jpatest;
-
-import com.junshock.jpatest.domain.Member;
-import org.springframework.stereotype.Repository;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-
-@Repository
-public class MemberRepository {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    public Long save(Member member){
-        em.persist(member);
-        return member.getId();
-    }
-
-    public Member find(Long id){
-        return em.find(Member.class, id);
-    }
-}
