+++ src/main/java/com/junshock/jpatest/domain/repository/MemberRepository.java
@@ -1,0 +1,35 @@
+package com.junshock.jpatest.domain.repository;
+
+import com.junshock.jpatest.domain.Member;
+import org.springframework.stereotype.Repository;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import java.util.List;
+
+@Repository
+public class MemberRepository {
+
+    @PersistenceContext // Spring이 EntityManager를 만들어서 주입.
+    private EntityManager em;
+
+    public Long save(Member member) {
+        em.persist(member);
+        return member.getId();
+    }
+
+    public Member find(Long id) {
+        return em.find(Member.class, id);
+    }
+
+    public List<Member> findAll() {
+        return em.createQuery("select m from Member m", Member.class) // JSQL 테이블이 아닌 엔터티 객체를 대상으로 질의
+                .getResultList();
+    }
+
+    public List<Member> findByName(String name) {
+        return em.createQuery("select m from Member m where m.name = :name", Member.class)
+                .setParameter("name", name)
+                .getResultList();
+    }
+}
